requirements_stage2.txt
llama-index-core>=0.12.0
llama-index-embeddings-ollama>=0.3.0
groq>=0.9.0
tiktoken>=0.7.0
lxml>=5.0.0
beautifulsoup4>=4.12.0
requests>=2.31.0
tenacity>=8.2.0
tqdm>=4.66.0
python-dotenv>=1.0.0
